pyproject.toml
[project]
name = "model-size-selection"
version = "0.1.0"
description = "Experimenting with model size selection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets<3.6.0",
    "openai>=1.105.0",
    "peft>=0.17.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.1",
    "mlflow>=2.16.2",
    "ipykernel>=6.30.1",
    "sentencepiece>=0.2.0",
    "tiktoken>=0.7.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]
